update: more general Declaration type (again)
The type of Declaration is now more general. The original idea was to
expose Strategies and Strategy declarations only. The drawback is that
one can not update an existing Strategy. This is very impractical for
processors that take several arguments, like decompose or
interpretations.

Now one can provide processor declarations together with update
functions. Processor declarations can be lifted to strategy declarations
with liftP. So it is possible to define also default strategies in terms
of the processor declaration.

The only drawback is now that processors have to be lifted explicitly to
strategies. This is ok. In practice there are just a few
processors where update functions are really practical.