add: solver command to configuration
Extend configuration and runtime state with a 'solver' component. The
Strategy type now also allows to update the runtime state. So, we can
set a default (solver,args) pair in the configuration and update it
using a strategy combinator 'withSolver'. The (solver,args) pair are
just strings. How they are actually processed depends on the processors.